🚚 move state type to new file
