+++ modules/client/src/types.ts
@@ -52,11 +52,12 @@ export {
   DepositParameters,
   GetChannelResponse,
   GetConfigResponse,
-  IConnextClient,
   IChannelProvider,
+  IConnextClient,
   INodeApiClient,
   InternalClientOptions,
   IRpcConnection,
+  IStoreService,
   KeyGen,
   JsonRpcRequest,
   LinkedTransferParameters,
