+++ modules/types/src/cf/data-types.ts
@@ -83,20 +83,6 @@ export type SingleAssetTwoPartyIntermediaryAgreement = {
   tokenAddress: string;
 };
 
-export type StateChannelJSON = {
-  readonly multisigAddress: string;
-  readonly proxyFactoryAddress: string;
-  readonly userNeuteredExtendedKeys: string[];
-  readonly proposedAppInstances: [string, AppInstanceProposal][];
-  readonly appInstances: [string, AppInstanceJson][];
-  readonly singleAssetTwoPartyIntermediaryAgreements: [
-    string,
-    SingleAssetTwoPartyIntermediaryAgreement
-  ][];
-  readonly freeBalanceAppInstance: AppInstanceJson | undefined;
-  readonly monotonicNumProposedApps: number;
-};
-
 export type AppInstanceInfo = {
   identityHash: string;
   appDefinition: string;
