+++ modules/types/src/cf/index.ts
@@ -19,7 +19,6 @@ import {
   SingleAssetTwoPartyCoinTransferInterpreterParams,
   singleAssetTwoPartyCoinTransferInterpreterParamsEncoding,
   SingleAssetTwoPartyIntermediaryAgreement,
-  StateChannelJSON,
   TwoPartyFixedOutcome,
   TwoPartyFixedOutcomeInterpreterParams,
   twoPartyFixedOutcomeInterpreterParamsEncoding,
@@ -90,7 +89,6 @@ export {
   SingleAssetTwoPartyIntermediaryAgreement,
   ContractABI,
   SolidityValueType,
-  StateChannelJSON,
   IRpcNodeProvider,
   CFCoreTypes,
   SignedStateHashUpdate,
