+++ modules/types/src/index.ts
@@ -9,4 +9,5 @@ export * from "./inputs";
 export * from "./messaging";
 export * from "./node";
 export * from "./pisaClient";
+export * from "./state";
 export * from "./store";
