+++ modules/types/src/store.ts
@@ -1,4 +1,10 @@
-import { StorePair } from "./channelProvider";
+export const ConnextNodeStorePrefix = "INDRA_NODE_CF_CORE";
+export const ConnextClientStorePrefix = "INDRA_CLIENT_CF_CORE";
+
+export type StorePair = {
+  path: string;
+  value: any;
+};
 
 export type InitCallback = (data: AsyncStorageData) => void;
 
@@ -38,4 +44,26 @@ export interface IBackupServiceAPI {
 export interface FileStorageOptions {
   fileExt?: string;
   fileDir?: string;
+}
+
+/**
+ * An interface for a stateful storage service with an API very similar to Firebase's API.
+ * Values are addressed by paths, which are separated by the forward slash separator `/`.
+ * `get` must return values whose paths have prefixes that match the provided path,
+ * keyed by the remaining path.
+ * `set` allows multiple values and paths to be atomically set. In Firebase, passing `null`
+ * as `value` deletes the entry at the given prefix, and passing objects with null subvalues
+ * deletes entries at the path extended by the subvalue's path within the object. `set` must
+ * have the same behaviour if the `allowDelete` flag is passed; otherwise, any null values or
+ * subvalues throws an error.
+ */
+export interface IStoreService {
+  get(path: string): Promise<any>;
+  set(pairs: { path: string; value: any }[], allowDelete?: Boolean): Promise<void>;
+  reset?(): Promise<void>;
+}
+
+export interface Store extends IStoreService {
+  set(pairs: StorePair[], shouldBackup?: Boolean): Promise<void>;
+  restore(): Promise<StorePair[]>;
 }
